src/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
pydantic==1.9.1
python-dotenv==0.20.0
sqlalchemy==1.4.39
psycopg2==2.9.3
alembic==1.8.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5